rename service folder to trident
