+++ pkg/trident/deploy_trident.go
@@ -1,4 +1,4 @@
-package services
+package trident
 
 import (
 	"context"
