+++ pkg/trident/user.go
@@ -1,4 +1,4 @@
-package services
+package trident
 
 import (
 	"context"
